Separated role entity client methods from properties...
...to bring them in line with other entities.
