+++ src/Disqord.Rest/Entities/Core/Guild/IRole.Client.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Disqord
+{
+    public partial interface IRole : ISnowflakeEntity, IMentionable, IDeletable
+    {
+        Task ModifyAsync(Action<ModifyRoleProperties> action, RestRequestOptions options = null);
+    }
+}
