+++ src/Disqord.Rest/Entities/Rest/Guild/RestRole.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Threading.Tasks;
-using Disqord.Models;
+﻿using Disqord.Models;
 
 namespace Disqord.Rest
 {
-    public sealed class RestRole : RestSnowflakeEntity, IRole
+    public sealed partial class RestRole : RestSnowflakeEntity, IRole
     {
         public string Name { get; private set; }
 
@@ -43,15 +41,6 @@ namespace Disqord.Rest
             IsManaged = model.Managed;
             IsMentionable = model.Mentionable;
         }
-
-        public async Task ModifyAsync(Action<ModifyRoleProperties> action, RestRequestOptions options = null)
-        {
-            var model = await Client.InternalModifyRoleAsync(GuildId, Id, action, options).ConfigureAwait(false);
-            Update(model);
-        }
-
-        public Task DeleteAsync(RestRequestOptions options = null)
-            => Client.DeleteRoleAsync(GuildId, Id, options);
 
         public override string ToString()
             => Name;
